requirements-production.txt
BeautifulSoup==3.2.0
Django==1.6.1
Markdown==2.4
Pillow==2.2.2
South==0.8.3
boto==2.8.0
celery==3.0.17
celerymon==0.4.0
django-celery==3.0.11
django-celery-with-redis==3.0
django-compressor==1.3
django-taggit==0.11.2
django-treebeard==2.0b1
psycopg2==2.2.1
scrubber==1.6.1
timelib==0.2.1
django-gevent-deploy==1.0
gevent==0.13.8
unidecode==0.04.14
bcrypt==1.0.2
